% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_project}
\alias{get_project}
\title{get_project}
\usage{
get_project(id = NULL, title = NULL)
}
\arguments{
\item{id}{project id}

\item{title}{project title}
}
\description{
get project information of a project specified by id or title
}
\examples{
get_project(id = 1)
}
